back-app/requirements.txt
fastapi[all]
python-dotenv
playwright
requests
uvicorn
undetected-chromedriver
mysql-connector-python
crawl4ai
bcrypt